more detailed volspace
Take this or leave this. @satra was not impressed with `volspace` today and I thought we required more explanation. I'm also adding, and we can totally drop this. The idea that the corresponding `json` tag for `volspace` should be `SpatialReferenceVolume` instead of just `SpatialReference`. It just seemed cleaner than having both `space-` and `volspace` correspond to the same, one, metadata key.